+++ src/subscriptions/v2/subscriptions-v2.test.ts
@@ -1,6 +1,14 @@
 import prismaClient from '../../../test-utils/prisma/prisma-client';
 import Salable from '../..';
-import { PaginatedSubscription, Invoice, Plan, Subscription, PaginatedSubscriptionInvoice, Version } from '../../types';
+import {
+  PaginatedSubscription,
+  Invoice,
+  Plan,
+  Subscription,
+  PaginatedSubscriptionInvoice,
+  Version,
+  PaginatedLicenses, Capability, License, SeatActionType
+} from '../../types';
 import getEndTime from '../../../test-utils/helpers/get-end-time';
 import { v4 as uuidv4 } from 'uuid';
 import { testUuids } from '../../../test-utils/scripts/create-test-data';
@@ -17,7 +25,7 @@ const licenseThreeUuid = uuidv4();
 const perSeatBasicLicenseUuids = [uuidv4(), uuidv4(), uuidv4(), uuidv4(), uuidv4(), uuidv4()];
 const testGrantee = '123456';
 const testEmail = 'tester@domain.com';
-const owner = 'subscription-owner'
+const owner = 'subscription-owner';
 
 describe('Subscriptions V2 Tests', () => {
   const apiKey = testUuids.devApiKeyV2;
@@ -82,7 +90,7 @@ describe('Subscriptions V2 Tests', () => {
       expand: ['plan'],
       sort: 'desc',
       productUuid: testUuids.productUuid,
-      planUuid: testUuids.paidPlanTwoUuid
+      planUuid: testUuids.paidPlanTwoUuid,
     });
 
     expect(dataWithSearchParams).toEqual({
@@ -98,7 +106,7 @@ describe('Subscriptions V2 Tests', () => {
           productUuid: testUuids.productUuid,
           plan: {
             ...planSchema,
-            uuid: testUuids.paidPlanTwoUuid
+            uuid: testUuids.paidPlanTwoUuid,
           },
         }),
       ]),
@@ -107,7 +115,7 @@ describe('Subscriptions V2 Tests', () => {
 
   it('getAll (w/ search params owner): Should successfully fetch subscriptions', async () => {
     const dataWithSearchParams = await salable.subscriptions.getAll({
-      owner: 'different-owner'
+      owner: 'different-owner',
     });
 
     expect(dataWithSearchParams).toEqual({
@@ -116,7 +124,21 @@ describe('Subscriptions V2 Tests', () => {
       data: expect.arrayContaining([{ ...subscriptionSchema }]),
     });
     expect(dataWithSearchParams.data.length).toEqual(1);
-    expect(dataWithSearchParams.data).toEqual([{...subscriptionSchema, owner: 'different-owner'}]);
+    expect(dataWithSearchParams.data).toEqual([{ ...subscriptionSchema, owner: 'different-owner' }]);
+  });
+
+  it('getSeats: Should successfully fetch a subscription\'s seats', async () => {
+    const data = await salable.subscriptions.getSeats(perSeatSubscriptionUuid);
+    expect(data).toEqual(paginatedLicensesSchema);
+  });
+
+  it('getSeatCount: Should successfully fetch a subscription\'s seat count', async () => {
+    const data = await salable.subscriptions.getSeatCount(perSeatSubscriptionUuid);
+    expect(data).toEqual({
+      count: expect.any(Number),
+      assigned: expect.any(Number),
+      unassigned: expect.any(Number),
+    });
   });
 
   it('getOne: Should successfully fetch the specified subscription', async () => {
@@ -140,10 +162,7 @@ describe('Subscriptions V2 Tests', () => {
   });
 
   it('getInvoices (w/ search params): Should successfully fetch a subscriptions invoices', async () => {
-    const data = await salable.subscriptions.getInvoices(
-      basicSubscriptionUuid,
-      { take: 1 }
-    );
+    const data = await salable.subscriptions.getInvoices(basicSubscriptionUuid, { take: 1 });
 
     expect(data).toEqual(stripeInvoiceSchema);
     expect(data.data.length).toEqual(1);
@@ -186,6 +205,16 @@ describe('Subscriptions V2 Tests', () => {
     expect(data).toBeUndefined();
   });
 
+  it('manageSeats: Should successfully perform multiple seat actions', async () => {
+    const data = await salable.subscriptions.manageSeats(perSeatSubscriptionUuid, [
+      { type: SeatActionType.assign, granteeId: 'assign_grantee_id' },
+      { type: SeatActionType.unassign, granteeId: 'userId_0' },
+      { type: SeatActionType.replace, granteeId: 'userId_1', newGranteeId: 'replace_grantee_id' },
+    ]);
+
+    expect(data).toBeUndefined();
+  });
+
   it('addSeats: Should successfully add seats to the subscription', async () => {
     const data = await salable.subscriptions.addSeats(perSeatSubscriptionUuid, {
       increment: 1,
@@ -216,6 +245,42 @@ describe('Subscriptions V2 Tests', () => {
     expect(data).toBeUndefined();
   });
 });
+
+const licenseCapabilitySchema: Capability = {
+  uuid: expect.any(String),
+  productUuid: expect.any(String),
+  name: expect.any(String),
+  status: expect.any(String),
+  description: expect.toBeOneOf([expect.any(String), null]),
+  updatedAt: expect.any(String),
+};
+
+const licenseSchema: License = {
+  uuid: expect.any(String),
+  name: expect.toBeOneOf([expect.any(String), null]),
+  email: expect.toBeOneOf([expect.any(String), null]),
+  subscriptionUuid: expect.toBeOneOf([expect.any(String), null]),
+  status: expect.toBeOneOf(['ACTIVE', 'CANCELED', 'EVALUATION', 'SCHEDULED', 'TRIALING', 'INACTIVE']),
+  granteeId: expect.toBeOneOf([expect.any(String), null]),
+  paymentService: expect.toBeOneOf(['ad-hoc', 'salable', 'stripe_existing']),
+  purchaser: expect.any(String),
+  type: expect.toBeOneOf(['licensed', 'metered', 'perSeat', 'customId', 'user']),
+  productUuid: expect.any(String),
+  planUuid: expect.any(String),
+  capabilities: expect.arrayContaining([licenseCapabilitySchema]),
+  metadata: expect.toBeOneOf([expect.anything(), null]),
+  startTime: expect.any(String),
+  endTime: expect.any(String),
+  updatedAt: expect.any(String),
+  isTest: expect.any(Boolean),
+  cancelAtPeriodEnd: expect.any(Boolean),
+};
+
+const paginatedLicensesSchema: PaginatedLicenses = {
+  first: expect.toBeOneOf([expect.any(String), null]),
+  last: expect.toBeOneOf([expect.any(String), null]),
+  data: expect.arrayContaining([licenseSchema]),
+};
 
 const planSchema: Plan = {
   uuid: expect.any(String),
@@ -283,6 +348,7 @@ const invoiceSchema: Invoice = {
   amount_due: expect.any(Number),
   amount_paid: expect.any(Number),
   amount_remaining: expect.any(Number),
+  amount_overpaid: expect.any(Number),
   amount_shipping: expect.any(Number),
   application: expect.toBeOneOf([expect.any(String), null]),
   application_fee_amount: expect.toBeOneOf([expect.any(Number), null]),
@@ -338,7 +404,7 @@ const invoiceSchema: Invoice = {
   quote: expect.toBeOneOf([expect.any(String), null]),
   receipt_number: expect.toBeOneOf([expect.any(String), null]),
   rendering: expect.toBeOneOf([expect.toBeObject(), null]),
-  rendering_options: expect.toBeOneOf([expect.toBeObject(), null]),
+  rendering_options: expect.toBeOneOf([expect.toBeObject(), undefined]),
   shipping_cost: expect.toBeOneOf([expect.toBeObject(), null]),
   shipping_details: expect.toBeOneOf([expect.toBeObject(), null]),
   starting_balance: expect.any(Number),
@@ -351,11 +417,13 @@ const invoiceSchema: Invoice = {
   subtotal_excluding_tax: expect.any(Number),
   tax: expect.toBeOneOf([expect.any(Number), null]),
   test_clock: expect.toBeOneOf([expect.any(String), null]),
+  parent: expect.toBeObject(),
   total: expect.any(Number),
   total_discount_amounts: expect.toBeOneOf([expect.toBeArray(), null]),
   total_excluding_tax: expect.any(Number),
   total_pretax_credit_amounts: expect.toBeOneOf([expect.toBeArray(), null]),
   total_tax_amounts: expect.toBeArray(),
+  total_taxes: expect.toBeArray(),
   transfer_data: expect.toBeOneOf([expect.toBeObject(), null]),
   webhooks_delivered_at: expect.toBeOneOf([expect.any(Number), null]),
 };
@@ -562,7 +630,7 @@ const generateTestData = async () => {
       createdAt: new Date(),
       updatedAt: new Date(),
       expiryDate: new Date(Date.now() + 31536000000),
-    }
+    },
   });
 
   await prismaClient.subscription.create({
@@ -586,9 +654,9 @@ const generateTestData = async () => {
 
   await prismaClient.subscription.create({
     data: {
+      uuid: perSeatSubscriptionUuid,
       lineItemIds: [stripeEnvs.perSeatBasicSubscriptionLineItemId],
       paymentIntegrationSubscriptionId: stripeEnvs.perSeatBasicSubscriptionId,
-      uuid: perSeatSubscriptionUuid,
       email: testEmail,
       owner,
       type: 'salable',